+++ src/commands/index.ts
@@ -1,2 +1,3 @@
 export { fetchCommand } from "./fetch";
 export { storeCommand } from "./store";
+export { getCommand, getsCommand } from "./get";